Updated knowledgebase and sfia_v8
